[#1] Refactor : 파일 리팩토링 및 바텀바 수정
