+++ app/src/main/java/com/example/mindstone/ui/mypage/Account1Fragment.kt
@@ -1,14 +1,14 @@
-package com.example.mindstone
+package com.example.mindstone.ui.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.mindstone.R
 import com.example.mindstone.databinding.FragmentAccount1Binding
 
 class Account1Fragment : Fragment() {
