+++ app/src/main/java/com/example/mindstone/ui/mypage/Account1OptionAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.mindstone
+package com.example.mindstone.ui.mypage
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
