+++ app/src/main/java/com/example/mindstone/ui/mypage/HsettingFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mindstone
+package com.example.mindstone.ui.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.mindstone.databinding.FragmentHsettingBinding
+import com.example.mindstone.domain.entity.Habit
 
 class HsettingFragment : Fragment() {
