+++ app/src/main/java/com/example/mindstone/ui/mypage/SettingFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mindstone
+package com.example.mindstone.ui.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.mindstone.TimeListAdapter
 import com.example.mindstone.databinding.DialogTimePickerBinding
 import com.example.mindstone.databinding.FragmentSettingBinding
