+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -13,10 +13,6 @@
         android:icon="@drawable/ic_home"
         android:title="홈"/>
     <item
-        android:id="@+id/nav_friends"
-        android:icon="@drawable/ic_friends"
-        android:title="친구목록"/>
-    <item
         android:id="@+id/nav_my"
         android:icon="@drawable/ic_my"
         android:title="마이페이지"/>
